complete

+++ proxy_server_with_cache.c
@@ -185,7 +185,8 @@ int handle_request(int clientSocket, ParsedRequest *request, char *tempReq)
 	{
 		bytes_send = send(clientSocket, buf, bytes_send, 0);
 		
-		for(int i=0;i<bytes_send/sizeof(char);i++){
+		for(size_t i = 0; i < bytes_send / sizeof(char); i++) {
+
 			temp_buffer[temp_buffer_index] = buf[i];
 			// printf("%c",buf[i]); // Response Printing
 			temp_buffer_index++;
@@ -269,7 +270,7 @@ void* thread_fn(void* socketNew)
 	
 	char *tempReq = (char*)malloc(strlen(buffer)*sizeof(char)+1);
     //tempReq, buffer both store the http request sent by client
-	for (int i = 0; i < strlen(buffer); i++)
+	for (size_t i = 0; i < strlen(buffer); i++)
 	{
 		tempReq[i] = buffer[i];
 	}
